src/test/java/br/com/pontointeligente/util/PasswordUtilsTest.java
package br.com.pontointeligente.util;

import org.junit.Assert;
import org.junit.Test;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;

public class PasswordUtilsTest {

    private static final String SENHA = "123456";
    private final BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();

    @Test
    public void testSenhaNula() throws Exception {
        Assert.assertNull(PasswordUtils.gerarByCrypt(null));
    }

    @Test
    public void testGerarHashSenha() throws Exception {
        String hash = PasswordUtils.gerarByCrypt(SENHA);
        Assert.assertTrue(bCryptPasswordEncoder.matches(SENHA,hash));
    }

}
